+++ src/repositories/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod boards;
 pub(crate) mod comments;
 pub(crate) mod elements;
+pub(crate) mod notifications;
 pub(crate) mod organizations;
 pub(crate) mod presence;
 pub(crate) mod realtime;